Create class BookingEntry and convert all arrays that contain booking data to use it
BookingEntry doesn't yet contain any useful methods, this will be the next
step.  Note that this code is currently PHP5 only (could be changed
later), because all methodes that take a BookingEntry have appropriate
type hinting in their signature, and because PHP4 can't iterate over an
object, which is needed in a few places.

Also, function documentation for functions that take a BookingEntry object
is not yet updated.